Cleanup App Check
Since it doesn't depend on bootstraping the Angular app, I've moved `setupAppCheck.ts` to a parent folder and outside of the `src` folder for the Angular target.
